feat(内存优化): 添加内存监控和自动调整功能
- 在数据加载器中实现实时内存监控和阈值检查
- 根据内存使用率自动调整worker数量
- 添加共享内存和内存映射配置
- 新增内存优化指南文档和测试脚本
- 修复数据加载器配置中的类型错误

+++ generate_data/pde_process/resolution_downsampler/resolution_downsampler.py
@@ -478,7 +478,7 @@ class DownsampledResolutionDataset(torch.utils.data.Dataset):
         return (
             torch.FloatTensor(input_flat),
             torch.FloatTensor(output_flat),
-            torch.tensor(idx, dtype=torch.float32)
+            torch.tensor(idx, dtype=torch.long)
         )
     
     def get_data_statistics(self):
